requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.14.0
django-crispy-forms
crispy-bootstrap5
Pillow==10.3.0